# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Church check-in and registration API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /forms:
    get:
      operationId: listForms
      tags: [forms]
      summary: List all registration forms
      responses:
        "200":
          description: List of forms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Form"
    post:
      operationId: createForm
      tags: [forms]
      summary: Create a new registration form
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FormInput"
      responses:
        "201":
          description: Created form
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Form"
...
  /forms/by-slug/{embedSlug}:
    get:
      operationId: getFormBySlug
      tags: [forms]
      summary: Get a public form by its embed slug (no auth required)
      parameters:
        - name: embedSlug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Form with questions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FormWithQuestions"
        "404":
          description: Not found
  /forms/{formId}:
    get:
      operationId: getForm
      tags: [forms]
      summary: Get a form with its questions
      parameters:
        - name: formId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Form with questions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FormWithQuestions"
        "404":
          description: Not found
    put:
      operationId: updateForm
      tags: [forms]
      summary: Update a form
      parameters:
        - name: formId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FormInput"
      responses:
        "200":
          description: Updated form
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Form"
    delete:
      operationId: deleteForm
      tags: [forms]
      summary: Delete a form
      parameters:
        - name: formId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Form:
      type: object
      required: [id, title, isActive, isPublic]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        isActive:
          type: boolean
        isPublic:
          type: boolean
        allowAdditionalPeople:
          type: boolean
        embedSlug:
          type: string
        submissionCount:
          type: integer
        createdAt:
          type: string
    FormInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
        description:
          type: string
        isActive:
          type: boolean
        isPublic:
          type: boolean
        allowAdditionalPeople:
          type: boolean
    FormWithQuestions:
      type: object
      required: [id, title, isActive, isPublic, questions]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        isActive:
          type: boolean
        isPublic:
          type: boolean
        allowAdditionalPeople:
          type: boolean
        embedSlug:
          type: string
        submissionCount:
          type: integer
        createdAt:
          type: string
        registrationType:
          type: ["string", "null"]
        eventId:
          type: ["integer", "null"]
        roomAssignmentMode:
          type: ["string", "null"]
        questions:
          type: array
          items:
            $ref: "#/components/schemas/Question"
        formFields:
          type: array
          items:
            $ref: "#/components/schemas/FormField"
    Question:
      type: object
      required: [id, formId, label, type, required, order]
      properties:
        id:
          type: integer
        formId:
          type: integer
        label:
          type: string
        type:
          type: string
          enum: [text, textarea, select, multiselect, checkbox, date, phone, email, number]
        required:
          type: boolean
        order:
          type: integer
        placeholder:
          type: ["string", "null"]
        options:
          type: ["string", "null"]
        isChildField:
          type: boolean
        fieldKey:
          type: ["string", "null"]
...
    FormField:
      type: object
      required: [id, formId, fieldKind, label, fieldType, required, sortOrder]
      properties:
        id:
          type: integer
        formId:
          type: integer
        fieldKind:
          type: string
          enum: [system, custom]
        systemKey:
          type: ["string", "null"]
        label:
          type: string
        fieldType:
          type: string
          enum: [text, textarea, select, multiselect, checkbox, date, phone, email, number]
        placeholder:
          type: ["string", "null"]
        required:
          type: boolean
        sortOrder:
          type: integer
        options:
          type: ["string", "null"]
        sectionKey:
          type: ["string", "null"]
...