# lib/api-zod/src/generated/types/formField.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Church check-in and registration API
 * OpenAPI spec version: 0.1.0
 */
import type { FormFieldFieldKind } from './formFieldFieldKind';
import type { FormFieldFieldType } from './formFieldFieldType';

export interface FormField {
  id: number;
  formId: number;
  fieldKind: FormFieldFieldKind;
  /** @nullable */
  systemKey?: string | null;
  label: string;
  fieldType: FormFieldFieldType;
  /** @nullable */
  placeholder?: string | null;
  required: boolean;
  sortOrder: number;
  /** @nullable */
  options?: string | null;
  /** @nullable */
  sectionKey?: string | null;
}